feat(homesection): show welcome message
